+++ test/fixtures/leaseOwner.ts
@@ -1,0 +1,3 @@
+import { acquireRepositoryLease } from '../../src/state/lease';
+
+void acquireRepositoryLease(process.argv[2]).then(() => process.send?.('ready'));